migration adds data to branches, user, kitstate, logtypes and settings

+++ app/database/migrations/2015_03_18_164155_create_KitContents_table.php
@@ -17,7 +17,7 @@ class CreateKitContentsTable extends Migration {
 			$table->increments('ID');
 			$table->integer('KitID');
 			$table->string('SerialNumber');
-			$table->tinyInteger('Damaged');
+			$table->integer('Damaged');
 			$table->tinyInteger('Missing');
 			$table->timestamps();
 		});
